Moves html-to-jsx out
